+++ packages/global.d.ts
@@ -1,0 +1,17 @@
+// Global compile-time constants
+declare let __DEV__: boolean;
+declare let __TEST__: boolean;
+declare let __BROWSER__: boolean;
+declare let __GLOBAL__: boolean;
+declare let __ESM_BUNDLER__: boolean;
+declare let __ESM_BROWSER__: boolean;
+declare let __NODE_JS__: boolean;
+declare let __COMMIT__: string;
+declare let __VERSION__: string;
+
+declare module "*.png" {
+  const value: any;
+  export default value;
+}
+
+