+++ packages/qms-kit/src/QmsKit.ts
@@ -1,4 +1,5 @@
-import WebMap, {
+import {
+  WebMap,
   StarterKit,
   Type,
   MainLayerAdapter,